Feat: Category parsing added
Now it is possible to make API calls to wikipedia to fetch the
categories that any wikipedia page belongs to. There is also support for
caching to optimise the number of calls to the wiki API.

TODO: Currently the cache is persistent only across 1 run. It would be
ideal to save and load the cache across multiple runs to further slower
runtimes.